明細能夠顯示，但是邏輯很奇怪(我可能不該join商品資料)

+++ MVC_Demo2/Models/ViewModel/HW_01_庫存盤點明細檔BasicViewModel.cs
@@ -16,10 +16,10 @@ namespace MVC_Demo2.Models.ViewModel
         public DateTime 日期 { get; set; }
 
         [Key]
-        public int 流水號 { get; set; }
+        public decimal 流水號 { get; set; }
 
         [Key]
-        public int 項次 { get; set; }
+        public decimal 項次 { get; set; }
 
         [CTRequired]
         public string 商品編號 { get; set; }
